# docs/openapi.yml
openapi: '3.0.3'
info:
  title: Tripost API
  version: '1.0'
  description: Travel SNS API for sharing trip experiences
servers:
  - url: https://mytripost.com
    description: Production server
  - url: http://localhost
    description: Development server (Laravel Sail)
paths:
...
  /auth/register:
    post:
      summary: Register a new user
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [displayid, email, password, password_confirmation]
              properties:
                displayid:
                  type: string
                  minLength: 5
                  maxLength: 50
                  description: Alphanumeric only, min 5 chars
                email:
                  type: string
                  format: email
                  maxLength: 255
                  pattern: '^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$'
                password:
                  type: string
                  minLength: 8
                  description: Must meet complexity requirements
                password_confirmation:
                  type: string
      responses:
        '302':
          description: Redirect to profile.edit after registration
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      summary: Login user
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                remember:
                  type: boolean
                  default: false
      responses:
        '302':
          description: Redirect to posts.index after login
        '422':
          description: Invalid credentials or validation error
...
  /profile/edit:
    get:
      summary: Show profile edit form
      tags: [Profile]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Profile edit form with user data and countries
  /profile/destroy:
    get:
      summary: Show account deletion confirmation page
      tags: [Profile]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Account deletion confirmation page

    post:
      summary: Delete own account (実装では DELETE だが OpenAPI では POST に変更)
      tags: [Profile]
      security:
        - cookieAuth: []
        - csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:
                  type: string
                _method:
                  type: string
                  enum: [DELETE]
                  description: Laravel method spoofing (to simulate DELETE)
      responses:
        '302':
          description: Redirect to /auth/register with success message
        '422':
          description: Validation error (e.g., incorrect password)
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: laravel_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-TOKEN
  schemas:
...
    Error:
      type: object
      properties:
        message: {type: string}
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string